restore goose.models module
