+++ liesel/goose/__init__.py
@@ -7,15 +7,10 @@ from .engine import Engine as Engine
 from .epoch import EpochConfig, EpochType
 from .gibbs import GibbsKernel
 from .hmc import HMCKernel
-from .interface import (
-    DataclassInterface,
-    DataClassModel,
-    DictInterface,
-    DictModel,
-    LieselInterface,
-)
+from .interface import DataclassInterface, DictInterface, LieselInterface
 from .iwls import IWLSKernel
 from .mh_kernel import MHKernel, MHProposal
+from .models import DataClassModel, DictModel
 from .nuts import NUTSKernel
 from .rw import RWKernel
 from .summary_m import Summary
